fix(notification): uplift state of notification count
